InterKosenCTF2020/matsushima2/challenge/requirements.txt
-i https://pypi.org/simple
click==7.1.2
flask==1.1.2
gevent==20.6.2
greenlet==0.4.16; platform_python_implementation == 'CPython'
itsdangerous==1.1.0
jinja2==2.11.2
markupsafe==1.1.1
pyjwt==1.7.1
uwsgi==2.0.19.1
werkzeug==1.0.1
zope.event==4.4
zope.interface==5.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
